+++ weed/util/wildcard/collection_filter_test.go
@@ -1,4 +1,4 @@
-package pluginworker
+package wildcard
 
 import "testing"
